pyproject.toml
[tool.poetry]
name = "ai-knowledge-base"
version = "0.1.0"
description = "AI-Powered Knowledge Base Search & Enrichment System"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
redis = "^5.0.1"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
pypdf2 = "^3.0.1"
python-docx = "^1.1.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.0"
openai = "^1.3.7"
tiktoken = "^0.5.2"
chromadb = "^0.4.18"
numpy = "^1.26.2"
pandas = "^2.1.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
tenacity = "^8.2.3"
asyncio-mqtt = "^0.16.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.1"
flake8 = "^6.1.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true